+++ config/config.go
@@ -3,8 +3,10 @@ package config
 // Config : A type that handles the configuration of the app
 type Config struct {
 	Vault struct {
-		Address string
-		Token   string
+		Address  string
+		Token    string
+		RoleId   string
+		SecretId string
 	}
 
 	State struct {